requirements.txt

asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.7.0
djangorestframework==3.12.2
pytz==2021.3
sqlparse==0.4.2